[구현] 신고하기, 아이템 디테일에 교환 요청 버튼 추가
