+++ navigation/Home.tsx
@@ -16,6 +16,7 @@ import Request from './HomeStack/Request';
 import ItemDeals from './ItemDeals';
 import ItemRequestDetail from './HomeStack/ItemRequestDetail';
 import ItemReceiveDetail from './HomeStack/ItemReceiveDetail';
+import Report from './HomeStack/Report';
 
 const Stack = createNativeStackNavigator();
 
@@ -93,6 +94,14 @@ export default function Home({ navigation }:HomeScreenProps) {
       />
       <Stack.Screen name="ItemRequestDetail" component={ItemRequestDetail} />
       <Stack.Screen name="ItemReceiveDetail" component={ItemReceiveDetail} />
+      <Stack.Screen
+        name="Report"
+        component={Report}
+        options={{
+        //   headerShown: false,
+          headerTitle: '문의하기',
+        }}
+      />
 
     </Stack.Navigator>
   );
