+++ navigation/HomeStack/ItemDetail.tsx
@@ -1,47 +1,27 @@
 import React from 'react';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import styled from 'styled-components/native';
 import Slide from './components/Slide';
 import {
   CommonText, InputColumn, Inputs, InputValue,
 } from '../ItemStack/utils';
 import { itemType } from '../../types';
-
-const Container = styled.ScrollView``;
-
-const ItemName = styled.Text`
-  font-size: 36;
-  font-weight:600;
-
-`;
-const ItemNameContainer = styled.View`
-  flex-direction: column;
-  /* justify-content: space-between; */
-  align-items: flex-start;
-  padding-bottom: 20px;
-  /* padding-left:10px; */
-`;
-
-const Deposit = styled.Text`
-  font-size: 20;
-  font-weight:600;
-`;
-
-const Description = styled.Text`
-  
-`;
-
-const DescriptionContainer = styled.View`
-  background-color: #eae9e9;
-  padding: 10px;
-  border-radius: 15px;
-
-`;
+import {
+  Button,
+  Container,
+  Deposit,
+  Description,
+  DescriptionContainer,
+  ItemName,
+  ItemNameContainer,
+  LikeButton,
+  ReportButton,
+} from './components/ItemDetailsComponents';
 
 export default function ItemDetail({ route }) {
   console.log(route.params);
   const { item } : {item:itemType} = route.params;
   const {
+    idx,
     name,
     description,
     deposit,
@@ -79,6 +59,8 @@ export default function ItemDetail({ route }) {
               }}
             />
           </InputColumn>
+          <LikeButton />
+          <ReportButton idx={idx} />
         </Inputs>
       </Container>
     </KeyboardAwareScrollView>
